#2532 disconnect FiltersFrequent/NotFrequent from the store #2532 move `parseDateToReadable` to Filters to easily import
